pyproject.toml
[tool.poetry]
name = "proraf"
version = "0.1.0"
description = "Sistema ProRAF - API de Rastreabilidade Agrícola"
authors = ["Rafael Nogueira <rafaelnogueira.aluno@unipampa.edu.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
factory-boy = "^3.3.0"
faker = "^20.1.0"
taskipy = "^1.12.2"
ruff = "^0.1.6"
httpx = "^0.25.2"

[tool.taskipy.tasks]
dev = "uvicorn proraf.app:app --reload --host 0.0.0.0 --port 8000"
test = "pytest -v -s --cov=proraf --cov-report=term-missing"
lint = "ruff check . --fix"
format = "ruff format ."
migrate = "alembic upgrade head"
makemigrations = "alembic revision --autogenerate -m"
init-db = "alembic upgrade head"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-v --cov=proraf"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"